# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Betayin Kino",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/movie/add": {
            "post": {
                "description": "Adds a new movie with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movie"
                ],
                "summary": "Add a new movie",
                "parameters": [
                    {
                        "description": "Movie Add Request",
                        "name": "addMovieReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/movieservice.AddMovieReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Response",
                        "schema": {
                            "$ref": "#/definitions/movieservice.AddMovieRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/movie/all": {
            "get": {
                "description": "Retrieves a list of all movies with optional filters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movie"
                ],
                "summary": "Get all movies",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by movie name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by studio",
                        "name": "studio",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by genre",
                        "name": "genre",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit the number of results",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset for pagination",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved list of movies",
                        "schema": {
                            "$ref": "#/definitions/movieservice.GetAllMoviesRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/movie/delete": {
            "delete": {
                "description": "Deletes a movie by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movie"
                ],
                "summary": "Delete a movie",
                "parameters": [
                    {
                        "description": "Movie Delete Request",
                        "name": "deleteMovieReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/movieservice.DeleteMovieReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Response",
                        "schema": {
                            "$ref": "#/definitions/movieservice.DeleteMovieRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/movie/update": {
            "put": {
                "description": "Updates the details of a movie by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movie"
                ],
                "summary": "Update movie details",
                "parameters": [
                    {
                        "description": "Movie Update Request",
                        "name": "updateMovieReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/movieservice.UpdateMovieReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Response",
                        "schema": {
                            "$ref": "#/definitions/movieservice.UpdateMovieRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/movie/{id}": {
            "get": {
                "description": "Retrieves details of a movie by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movie"
                ],
                "summary": "Get movie details by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Movie ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Response",
                        "schema": {
                            "$ref": "#/definitions/movieservice.GetMovieByIdRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "movieservice.AddMovieReq": {
            "type": "object",
            "properties": {
                "age_limit": {
                    "type": "integer"
                },
                "background_image_url": {
                    "type": "string"
                },
                "bio": {
                    "type": "string"
                },
                "genres": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/movieservice.Genres"
                    }
                },
                "language": {
                    "type": "string"
                },
                "movie_name": {
                    "type": "string"
                },
                "movie_url": {
                    "type": "string"
                },
                "season": {
                    "type": "integer"
                },
                "studio": {
                    "type": "string"
                }
            }
        },
        "movieservice.AddMovieRes": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "movieservice.DeleteMovieReq": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "movieservice.DeleteMovieRes": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean"
                }
            }
        },
        "movieservice.Genres": {
            "type": "object",
            "properties": {
                "genre": {
                    "type": "string"
                }
            }
        },
...
        "movieservice.GetAllMoviesRes": {
            "type": "object",
            "properties": {
                "movies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/movieservice.GetMovieByIdRes"
                    }
                },
                "total_count": {
                    "type": "integer"
                }
            }
        },
...
        "movieservice.GetMovieByIdRes": {
            "type": "object",
            "properties": {
                "age_limit": {
                    "type": "integer"
                },
                "background_image_url": {
                    "type": "string"
                },
                "bio": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "genres": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/movieservice.Genres"
                    }
                },
                "id": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "movie_name": {
                    "type": "string"
                },
                "movie_url": {
                    "type": "string"
                },
                "studio": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "movieservice.UpdateMovieReq": {
            "type": "object",
            "properties": {
                "age_limit": {
                    "type": "integer"
                },
                "background_image_url": {
                    "type": "string"
                },
                "bio": {
                    "type": "string"
                },
                "genres": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/movieservice.Genres"
                    }
                },
                "id": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "movie_name": {
                    "type": "string"
                },
                "movie_url": {
                    "type": "string"
                },
                "season": {
                    "type": "integer"
                },
                "studio": {
                    "type": "string"
                }
            }
        },
        "movieservice.UpdateMovieRes": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authourization",
            "in": "header"
        }
    }
}